perf inject: Fix dso->nsinfo refcounting
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 0967ebffe098157180a0bbd180ac90348c6e07d7
Author: Riccardo Mancini <rickyman7@gmail.com>
Date: Thu Jul 15 18:07:06 2021 +0200

description
===========
ASan reports a memory leak of nsinfo during the execution of:

  # perf test "31: Lookup mmap thread"

The leak is caused by a refcounted variable being replaced without
dropping the refcount.

This patch makes sure that the refcnt of nsinfo is decreased when a
refcounted variable is replaced with a new value.

Fixes: 27c9c3424fc217da ("perf inject: Add --buildid-all option")
[ Split from a larger patch ]

Signed-off-by: Michael Petlan <mpetlan@redhat.com>