perf probe: Fix dso->nsinfo refcounting
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit dedeb4be203b382ba7245d13079bc3b0f6d40c65
Author: Riccardo Mancini <rickyman7@gmail.com>
Date: Thu Jul 15 18:07:06 2021 +0200

description
===========
ASan reports a memory leak of nsinfo during the execution of:

 # perf test "31: Lookup mmap thread".

The leak is caused by a refcounted variable being replaced without
dropping the refcount.

This patch makes sure that the refcnt of nsinfo is decreased whenever
a refcounted variable is replaced with a new value.

Fixes: 544abd44c7064c8a ("perf probe: Allow placing uprobes in alternate namespaces.")
[ Split from a larger patch ]

Signed-off-by: Michael Petlan <mpetlan@redhat.com>