perf map: Fix dso->nsinfo refcounting
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 2d6b74baa7147251c30a46c4996e8cc224aa2dc5
Author: Riccardo Mancini <rickyman7@gmail.com>
Date: Thu Jul 15 18:07:06 2021 +0200

description
===========
ASan reports a memory leak of nsinfo during the execution of

  # perf test "31: Lookup mmap thread"

The leak is caused by a refcounted variable being replaced without
dropping the refcount.

This patch makes sure that the refcnt of nsinfo is decreased whenever a
refcounted variable is replaced with a new value.

Fixes: bf2e710b3cb8445c ("perf maps: Lookup maps in both intitial mountns and inner mountns.")
[ Split from a larger patch ]

Signed-off-by: Michael Petlan <mpetlan@redhat.com>