perf dso: Fix memory leak in dso__new_map()
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 581e295a0f6b5c2931d280259fbbfff56959faa9
Author: Riccardo Mancini <rickyman7@gmail.com>
Date: Thu Jul 15 18:07:11 2021 +0200

description
===========
ASan reports a memory leak when running:

  # perf test "65: maps__merge_in".

The causes of the leaks are two, this patch addresses only the first
one, which is related to dso__new_map().

The bug is that dso__new_map() creates a new dso but never decreases the
refcount it gets from creating it.

This patch adds the missing dso__put().

Fixes: d3a7c489c7fd2463 ("perf tools: Reference count struct dso")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>