Allocate bw_dwork as a pointer and introduce a backpointer in the work struct
Due to KABI concerns we are introducing the bw_dwork as a pointer and
introduce a new backpointer in the work struct to point back to the
bdi_writeback struct in wb_update_bandwidth_workfn. We leave the
previous commit with the the RH_KABI_BROKEN_INSERT in-order to maintain
bisectability.

Upstream-Status: RHEL-Only
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1990580
Signed-off-by: Nico Pache <npache@redhat.com>