mm/mmu_notifier: Add mmu_notifier versioning mechanism
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1946746
Upstream Status: RHEL only

It is hard to upgrade the mm/mmu_notifier code in RHEL8 because of the
need to support stable kABI of the following mmu_notifier functions:
 - mmu_notifier_register()
 - __mmu_notifier_register()
 - mmu_notifier_unregister()

This translates to the kABI stability of the mmu_notifier_ops structure.

Upstream, however, had changed the function prototypes of some of the
callback methods, namely invalidate_range_start() and invalidate_range_end().
As a result, we cannot pull in some of the upstream mmu_notifier patches
because of that limitation.

One way to overcome that limitation is by using method versioning, i.e.
by providing a second set of methods with the new function prototypes.
Third party vendor can then choose the method version that they want by,
for instance, defining the macro RH_MMU_NOTIFIER_V2 to use the new v2
methods instead of legacy v1 methods.

This patch provides the infrastructure to do that and allowing us to
pull in upstream mmu_notifier patches that we are not able to do before.
We do use up all the remaining free RH_KABI_RESERVED() entries in the
mmu_notifier_ops structure for now. However, we also add a version
number into the mmu_notifier_rh structure. We don't need to use that
version number currently. If we ever need to bump up the version number
further in the future and increasing the size of mmu_notifier_ops, we
can use the version number to determine the size of the mmu_notifier_ops
structure that is passed in.

Two new v2 mmu_notifier registration functions are now provided:
 - mmu_notifier_register_v2()
 - __mmu_notifier_register_v2()

They will be used when the RH_MMU_NOTIFIER_V2 macro is enabled. Their
purpose is for setting the right version number as well as making sure
that third party kernel modules that use the v2 mmu_notifier API cannot
be loaded into earlier RHEL8 kernels that don't support the v2 API.

Signed-off-by: Waiman Long <longman@redhat.com>