ACPI: PM: Do not turn off power resources in unknown state
Upstream Status: Linus v5.15
Testing Status: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=39199451
        s390x storage blktests failed, but that failed on a default build
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1997696
Conflicts:
	drivers/acpi/power.c - fixup failed because of the difference
in debug printing macros, but the intention to change the comparison
and remove the comments was obvious.

commit bc28368596436e6e81ffc48c815b8225d96121c0
Author: "Rafael J. Wysocki" <rafael.j.wysocki@intel.com>
Date: Fri, 15 Oct 2021 19:12:21 +0200

Commit 6381195ad7d0 ("ACPI: power: Rework turning off unused power
resources") caused power resources in unknown state with reference
counters equal to zero to be turned off too, but that caused issues
to appear in the field, so modify the code to only turn off power
resources that are known to be "on".

Link: https://lore.kernel.org/linux-acpi/6faf4b92-78d5-47a4-63df-cc2bab7769d0@molgen.mpg.de/
Fixes: 6381195ad7d0 ("ACPI: power: Rework turning off unused power resources")
Reported-by: Andreas K. Huettel <andreas.huettel@ur.de>
Tested-by: Andreas K. Huettel <andreas.huettel@ur.de>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: 5.14+ <stable@vger.kernel.org> # 5.14+

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>