perf/x86/intel/uncore: Add Sapphire Rapids server IIO support
Bugzilla: https://bugzilla.redhat.com/1837402

upstream
========
commit 3ba7095beaec1bace1b5864fa92b2b7a0eaadf38
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Wed Jun 30 14:08:27 2021 -0700

description
===========
The IIO stacks are responsible for managing the traffic between the PCI
Express* (PCIe*) domain and the mesh domain. The IIO PMON block is
situated near the IIO stacks traffic controller capturing the traffic
controller as well as the PCIe* root port information.

The layout of the control registers for a IIO uncore unit is a little
bit different from the generic one.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>