perf/x86/intel/uncore: Add Sapphire Rapids server M2M support
Bugzilla: https://bugzilla.redhat.com/1837402

upstream
========
commit f57191edaaeb01279a88ace1be5b7230bdd8c0ab
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Wed Jun 30 14:08:32 2021 -0700

description
===========
The M2M blocks manage the interface between the mesh (operating on both
the mesh and the SMI3 protocol) and the memory controllers.

The layout of the control registers for a M2M uncore unit is a little
 bit different from the generic one. So a specific format and ops are
required. Expose the common PCI ops which can be reused.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>