software node: Fix device_add_software_node()
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2032600
Upstream status: In Linux v5.12
Conflicts:
       include/linux/property.h - fixup failed because
of difference contexts, but the change was to rename
swnode to node in device_add_software_node so that is
what I did.

commit 2a92c90f2ecca4475d6050f2f938a1755a8954cc
Author: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Date: Mon, 1 Mar 2021 17:30:12 +0300

The function device_add_software_node() was meant to
register the node supplied to it, but only if that node
wasn't already registered. Right now the function attempts
to always register the node. That will cause a failure with
nodes that are already registered.

Fixing that by incrementing the reference count of the nodes
that have already been registered, and only registering the
new nodes. Also, clarifying the behaviour in the function
documentation.

Fixes: e68d0119e328 ("software node: Introduce device_add_software_node()")
Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Tested-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 2a92c90f2ecca4475d6050f2f938a1755a8954cc)
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>