ALSA: usb-audio: Handle error for the current selector gracefully
Currently we bail out when the device returns an error or an invalid
value for the current clock selector value via
uac_clock_selector_get_val().  But it's possible that the device is
really uninitialized and waits for the setup of the proper route at
first.

For handling such a case, this patch lets the driver dealing with the
error or the invalid error more gracefully, choosing the clock source
automatically instead.

    Link: https://lore.kernel.org/r/20210518152112.8016-3-tiwai@suse.de
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Takashi Iwai <tiwai@suse.de>
Date: Tue May 18 17:21:12 2021 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 481f17c41803985446fd12887b2c042f9c43b0d5)
Bugzilla: https://bugzilla.redhat.com/2003613