ALSA: usb-audio: scarlett2: Fix for loop increment in scarlett2_usb_get_config
Clang warns:

sound/usb/mixer_scarlett_gen2.c:1189:32: warning: expression result
unused [-Wunused-value]
                        for (i = 0; i < count; i++, (u16 *)buf++)
                                                    ^      ~~~~~
1 warning generated.

It appears the intention was to cast the void pointer to a u16 pointer
so that the data could be iterated through like an array of u16 values.
However, the cast happens after the increment because a cast is an
rvalue, whereas the post-increment operator only works on lvalues, so
the loop does not iterate as expected. This is not a bug in practice
because count is not greater than one at the moment but this could
change in the future so this should be fixed.

Replace the cast with a temporary variable of the proper type, which is
less error prone and fixes the iteration. Do the same thing for the
'u8 *' below this if block.

    Fixes: ac34df733d2d ("ALSA: usb-audio: scarlett2: Update get_config to do endian conversion")
    Link: https://github.com/ClangBuiltLinux/linux/issues/1408
    Acked-by: Geoffrey D. Bennett <g@b4.vu>
    Signed-off-by: Nathan Chancellor <nathan@kernel.org>
    Link: https://lore.kernel.org/r/20210627051202.1888250-1-nathan@kernel.org
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Nathan Chancellor <nathan@kernel.org>
Date: Sat Jun 26 22:12:03 2021 -0700

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 9a7b7ec3c6f2aedb99c9f39dd9cb63c2deff3437)
Bugzilla: https://bugzilla.redhat.com/2003613