ALSA: usb-audio: Line6 HX-Stomp XL USB_ID for 48k-fixed quirk
Adding the Line6 HX-Stomp XL USB_ID as it needs this fixed frequency
quirk as well.

The device is basically just the HX-Stomp with some more buttons on
the face.  I've done some recording with it after adding it, and it
seems to function properly with this fix.  The Midi features appear to
be working as well.

[ a coding style fix and patch reformat by tiwai ]

    Signed-off-by: Jason Ormes <skryking@gmail.com>
    Cc: <stable@vger.kernel.org>
    Link: https://lore.kernel.org/r/20211030200405.1358678-1-skryking@gmail.com
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Jason Ormes <skryking@gmail.com>
Date: Sat Oct 30 15:04:05 2021 -0500

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 8f27b689066113a3e579d4df171c980c54368c4e)
Bugzilla: https://bugzilla.redhat.com/2003613