ALSA: hda: Drop workaround for a hang at shutdown again
The commit 0165c4e19f6e ("ALSA: hda: Fix hang during shutdown due to
link reset") modified the shutdown callback of the HD-audio controller
for working around a hang.  Meanwhile, the actual culprit of the hang
was identified to be the leftover active codecs that may interfere
with the powered down controller somehow, but we took a minimal fix
approach for 5.14, and that was the commit above.

Now, since the codec drivers go runtime-suspend at shutdown for 5.15,
we can revert the change and make sure that the full runtime-suspend
is performed at shutdown of HD-audio controller again.  This patch
essentially reverts the commit above to restore the behavior.

    BugLink: https://bugzilla.kernel.org/show_bug.cgi?id=214045
    Link: https://lore.kernel.org/r/20210817075630.7115-1-tiwai@suse.de
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Takashi Iwai <tiwai@suse.de>
Date: Tue Aug 17 09:56:30 2021 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 8fc8e903156f42c66245838441d03607e9067381)
Bugzilla: https://bugzilla.redhat.com/2003613