ASoC: rt715-sdw: use first_hw_init flag on resume
The intent of the status check on resume was to verify if a SoundWire
peripheral reported ATTACHED before waiting for the initialization to
complete. This is required to avoid timeouts that will happen with
'ghost' devices that are exposed in the platform firmware but are not
populated in hardware.

Unfortunately we used 'hw_init' instead of 'first_hw_init'. Due to
another error, the resume operation never timed out, but the volume
settings were not properly restored.

    BugLink: https://github.com/thesofproject/linux/issues/2908
    BugLink: https://github.com/thesofproject/linux/issues/2637
    Fixes: d1ede0641b05e ('ASoC: rt715: add RT715 codec driver')
    Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
    Reviewed-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
    Reviewed-by: Bard Liao <bard.liao@intel.com>
    Link: https://lore.kernel.org/r/20210607222239.582139-11-pierre-louis.bossart@linux.intel.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Date: Mon Jun 7 17:22:34 2021 -0500

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit dbc07517ab173688ef11234d1099bc1e24e4f14b)
Bugzilla: https://bugzilla.redhat.com/2003613