ASoC: soc-topology: Move template info print soc_tplg_dapm_widget_create()
A DAPM widget now can have different types of controls, it is no longer
correct to print the type as it is just the type of the first control.

Move it after the loop where we create the controls and print the number
of the control types.

    Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
    Reviewed-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
    Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
    Reviewed-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
    Link: https://lore.kernel.org/r/20210902112301.22657-1-peter.ujfalusi@linux.intel.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Date: Thu Sep 2 14:23:01 2021 +0300

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 8facf84bcf575e3217a15cefcc867db15dca4781)
Bugzilla: https://bugzilla.redhat.com/2003613