ASoC: cs42l42: Always configure both ASP TX channels
An I2S frame always has two slots (left and right) even when sending
mono. The right channel (channel 2) of ASP TX will always have the
same bit width as the left channel and will always be on the high
phase of LRCLK.

The previous implementation always passed the field masks for both
channels to snd_soc_component_update_bits() but for mono the written value
only contained the settings for channel 1. The result was that for mono
channel 2 was set to 8-bit (which is an invalid configuration) with both
channels on the low phase of LRCLK.

    Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
    Fixes: 585e7079de0e ("ASoC: cs42l42: Add Capture Support")
    Link: https://lore.kernel.org/r/20211015133619.4698-3-rf@opensource.cirrus.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Richard Fitzgerald <rf@opensource.cirrus.com>
Date: Fri Oct 15 14:36:05 2021 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 6e6825801ab926360f7f4f2dbcfd107d5ab8f025)
Bugzilla: https://bugzilla.redhat.com/2003613