ASoC: cs42l42: free_irq() before powering-down on probe() fail
Relying on devm to free the irq handler on probe failure leaves a
small window of opportunity for an interrupt to become pending and
then the handler to run after the chip has been reset and powered
off.

For safety cs42l42_probe() should free the irq in the error path.
As the irq is now disabled by the driver in probe() and remove()
there is no point allocating it as a devres-managed item, so
convert to plain non-devres.

    Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
    Link: https://lore.kernel.org/r/20211026125722.10220-3-rf@opensource.cirrus.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Richard Fitzgerald <rf@opensource.cirrus.com>
Date: Tue Oct 26 13:57:22 2021 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit a10148a8cf561d728c0f57994330b2da1df35577)
Bugzilla: https://bugzilla.redhat.com/2003613